+++ src/screens/SignInPassword/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useSignInPassword} from './useSignInPassword'
