+++ src/screens/SignUpForm/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useSignUpForm} from './useSignUpForm'
