+++ src/screens/SignUpForm/hooks/useSignUpForm.tsx
@@ -1,0 +1,46 @@
+import {useState, useCallback} from 'react';
+import {useDispatch} from 'react-redux';
+import {useRequestLoading, useTogglePasswordVisibility} from 'core/hooks';
+import {
+  SignUpFormScreenRouteProps,
+  SignUpFormScreenNavigationProps,
+} from '../types';
+import {useNavigation, useRoute} from '@react-navigation/native';
+import {signUpRequest} from 'core/reducers';
+
+export const useSignUpForm = () => {
+  const dispatch = useDispatch();
+  const [password, setPassword] = useState('');
+
+  const {
+    params: {email},
+  } = useRoute<SignUpFormScreenRouteProps>();
+  const navigation = useNavigation<SignUpFormScreenNavigationProps>();
+
+  const navigateToEmailValidation = useCallback(() => {
+    navigation.navigate('CodeVerification', {email, isSignUp: true});
+  }, [email, navigation]);
+
+  const signUp = useCallback(() => {
+    dispatch(signUpRequest({email, password}));
+  }, [dispatch, email, password]);
+
+  const {passwordVisibility, rightIcon, handlePasswordVisibility} =
+    useTogglePasswordVisibility('eye');
+
+  const {loading} = useRequestLoading(signUpRequest);
+
+  return {
+    dispatch,
+    loading,
+    email,
+    password,
+    setPassword,
+    navigation,
+    passwordVisibility,
+    rightIcon,
+    handlePasswordVisibility,
+    navigateToEmailValidation,
+    signUp,
+  };
+};
